+++ R/summary.R
@@ -1,6 +1,6 @@
 #' Display the summary of DynForest
 #'
-#' @param object \code{DynForest} or \code{DynForest_OOB} object
+#' @param object \code{DynForest} or \code{DynForestOOB} object
 #' @param ... Optional parameters to be passed to the low level function
 #'
 #' @return Return some information about the random forest
@@ -138,12 +138,98 @@ summary.DynForest <- function(object, ...){
 
   cat("----------------","\n")
 
-  # erreur out-of-bag
+  # computation time
+  cat("Computation time \n")
+  cat(paste0("\t","Number of cores used: ", object$ncores),"\n")
+  cat("\t")
+  print(object$comput.time)
+  cat("----------------","\n")
+
+}
+
+#' @rdname summary.DynForest
+#' @export
+summary.DynForestOOB <- function(object, ...){
+
+  if (object$type=="surv"){
+    if (length(object$causes)>1){
+      type <- "survival (competing risk)"
+      split.rule <- "Fine & Gray statistic test"
+    }else{
+      type <- "survival"
+      split.rule <- "Maximize log-rank statistic test"
+    }
+    oob.type <- "Integrated Brier Score"
+    leaf.stat <- "Cumulative incidence function"
+  }
+
+  if (object$type=="factor"){
+    type <- "categorical"
+    oob.type <- "Missclassification"
+    split.rule <- "Minimize weighted within-group Shannon entropy"
+    leaf.stat <- "Majority vote"
+  }
+
+  if (object$type=="numeric"){
+    type <- "continuous"
+    oob.type <- "Mean square error"
+    split.rule <- "Minimize weighted within-group variance"
+    leaf.stat <- "Mean"
+  }
+
+  ##############################
+
+  cat(paste0("DynForest executed for ", type, " outcome"),"\n")
+  cat(paste0("\t","Splitting rule: ", split.rule),"\n")
+  cat(paste0("\t","Out-of-bag error type: ", oob.type),"\n")
+  cat(paste0("\t","Leaf statistic: ", leaf.stat),"\n")
+  cat("----------------","\n")
+
+  # input
+  cat("Input","\n")
+  cat(paste0("\t","Number of subjects: ", length(unique(unlist(apply(object$rf, 2, FUN = function(x) x$idY))))),"\n")
+  cat(paste0("\t","Longitudinal: ", length(object$Inputs$Longitudinal), " predictor(s)"),"\n")
+  cat(paste0("\t","Numeric: ", length(object$Inputs$Numeric), " predictor(s)"),"\n")
+  cat(paste0("\t","Factor: ", length(object$Inputs$Factor), " predictor(s)"),"\n")
+  cat("----------------","\n")
+
+  # tuning parameters
+  cat("Tuning parameters","\n")
+  cat(paste0("\t","mtry: ", object$param$mtry),"\n")
+  cat(paste0("\t","nodesize: ", object$param$nodesize),"\n")
+  if (type%in%c("survival (competing risk)","survival")){
+    cat(paste0("\t","minsplit: ", object$param$minsplit),"\n")
+  }
+  cat(paste0("\t","ntree: ", object$param$ntree),"\n")
+  cat("----------------","\n")
+  cat("----------------","\n")
+
+  # DynForest summary
+  cat("DynForest summary","\n")
+  cat(paste0("\t","Average depth per tree: ",
+             round(mean(apply(object$rf, 2, FUN = function(x){
+               mean(x$V_split$depth[which(x$V_split$type=="Leaf")])}),
+               na.rm = T),2)),"\n")
+  cat(paste0("\t","Average number of leaves per tree: ",
+             round(mean(apply(object$rf, 2, FUN = function(x){
+               length(x$V_split$type[which(x$V_split$type=="Leaf")])}),
+               na.rm = T),2)),"\n")
+  cat(paste0("\t","Average number of subjects per leaf: ",
+             round(mean(apply(object$rf, 2, FUN = function(x){
+               mean(x$V_split$N[which(x$V_split$type=="Leaf")])}),
+               na.rm = T),2)),"\n")
+  if (type%in%c("survival (competing risk)","survival")){
+    cat(paste0("\t","Average number of events of interest per leaf: ",
+               round(mean(apply(object$rf, 2, FUN = function(x){
+                 mean(x$V_split$Nevent[which(x$V_split$type=="Leaf")])}),
+                 na.rm = T),2)),"\n")
+  }
+
+  cat("----------------","\n")
+
+  # out-of-bag error
   cat(paste0("Out-of-bag error based on ", oob.type),"\n")
-  cat(paste0("\t","Out-of-bag error: ",
-             ifelse(!is.null(object$oob.err),
-                    round(mean(object$oob.err, na.rm = T), 4),
-                    "Not computed!")),"\n")
+  cat(paste0("\t","Out-of-bag error: ", round(mean(object$oob.err, na.rm = T), 4)), "\n")
   cat("----------------","\n")
 
   # computation time